requirements.txt
streamlit==1.23.0
crewai
langchain_groq
crewai_tools
python-dotenv
ipython
pandas



